# docs/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "version": "1.0.0",
    "title": "Recipies app",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "description": "This is a documentation of recipes app"
  },
  "servers": [
    {
      "url": "https://backend-tasteorama.onrender.com",
      "description": "Production server"
    },
    {
      "url": "http://127.0.0.1:8080",
      "description": "Local development server"
    }
  ],
...
  "paths": {
    "/recipes/{id}": {
      "get": {
        "tags": [
          "Recipes"
        ],
        "summary": "Get recipe by id",
        "operationId": "getRecipeById",
        "description": "Get recipe by id",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "example": "65ca67e7ae7f10c88b598384"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get recipe by id",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully found recipe with id 65ca67e7ae7f10c88b598384!"
                    },
                    "data": {
                      "type": "object",
                      "$ref": "#/components/schemas/recipe"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "404": {
            "description": "Recipe not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 404
                    },
                    "message": {
                      "type": "string",
                      "example": "Recipes not found"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
    "/recipes": {
      "get": {
        "tags": [
          "Recipes"
        ],
        "summary": "Get list of Recipes",
        "operationId": "getRecipes",
        "description": "Get list of recipes with this endpoint",
        "security": [],
        "parameters": [
          {
            "in": "query",
            "name": "title",
            "schema": {
              "type": "string",
              "example": "Cake"
            },
            "description": "Partial match in recipe title"
          },
          {
            "in": "query",
            "name": "category",
            "schema": {
              "type": "string",
              "enum": [
                "Seafood",
                "Lamb",
                "Starter",
                "Chicken",
                "Beef",
                "Dessert",
                "Vegan",
                "Pork",
                "Vegetarian",
                "Miscellaneous",
                "Pasta",
                "Breakfast",
                "Side",
                "Goat",
                "Soup"
              ],
              "example": "Dessert"
            },
            "description": "Exact category match"
          },
          {
            "in": "query",
            "name": "ingredient",
            "schema": {
              "type": "string",
              "enum": [
                "Egg",
                "Pears",
                "Spring Onions",
                "Minced Beef",
                "Powdered Sugar",
                "Jalapeno",
                "Madras Paste",
                "Noodles"
              ],
              "example": "Egg"
            },
            "description": "Ingredient name to match (partial match allowed)"
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "example": 1
            },
            "description": "Page number (default is 1)"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "example": 12
            },
            "description": "Number of results per page (default is 10)"
          }
        ],
        "responses": {
          "200": {
            "description": "Recipes found successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Recipes retrieved successfully"
                    },
                    "data": {
                      "type": "object",
                      "required": [
                        "recipes",
                        "page",
                        "perPage",
                        "totalItems",
                        "totalPages",
                        "hasNextPage",
                        "hasPreviousPage"
                      ],
                      "properties": {
                        "recipes": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/recipe"
                          }
                        },
                        "page": {
                          "type": "integer",
                          "example": 1
                        },
                        "perPage": {
                          "type": "integer",
                          "example": 10
                        },
                        "totalItems": {
                          "type": "integer",
                          "example": 40
                        },
                        "totalPages": {
                          "type": "integer",
                          "example": 4
                        },
                        "hasNextPage": {
                          "type": "boolean",
                          "example": true
                        },
                        "hasPreviousPage": {
                          "type": "boolean",
                          "example": false
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Invalid query parameters",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 400
                    },
                    "message": {
                      "type": "string",
                      "example": "Invalid query parameters. Please check the title, category, or ingredient fields."
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found - No recipes match the given criteria",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 404
                    },
                    "message": {
                      "type": "string",
                      "example": "No recipes found for the provided criteria."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      },
      "post": {
        "tags": [
          "Recipes"
        ],
        "summary": "Create a new recipe",
        "operationId": "createRecipe",
        "description": "Create a new recipe",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "category",
                  "ingredient",
                  "ingredientAmount",
                  "instruction",
                  "decr",
                  "cookiesTime"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 64,
                    "example": "Chocolate Cake"
                  },
                  "category": {
                    "type": "string",
                    "enum": [
                      "Seafood",
                      "Lamb",
                      "Starter",
                      "Chicken",
                      "Beef",
                      "Dessert",
                      "Vegan",
                      "Pork",
                      "Vegetarian",
                      "Miscellaneous",
                      "Pasta",
                      "Breakfast",
                      "Side",
                      "Goat",
                      "Soup"
                    ],
                    "example": "Dessert"
                  },
                  "area": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 100,
                    "example": "Italian"
                  },
                  "decr": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200,
                    "example": "A rich chocolate cake with cream"
                  },
                  "cookiesTime": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 360,
                    "example": 45
                  },
                  "cals": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 10000,
                    "example": 350
                  },
                  "instruction": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 1200,
                    "example": "Mix everything and bake at 180°C for 40 min"
                  },
                  "thumb": {
                    "type": "string",
                    "format": "binary",
                    "description": "Recipe image file (optional, max size 2MB)"
                  },
                  "owner": {
                    "type": "string",
                    "description": "ID of the user (optional)",
                    "example": "685e469e90ec1af79417e1e0"
                  },
                  "ingredient": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "minItems": 1,
                    "description": "A list of ingredients for the recipe.",
                    "example": [
                      "Flour",
                      "Sugar",
                      "Butter"
                    ]
                  },
                  "ingredientAmount": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "minItems": 2,
                    "maxItems": 16,
                    "description": "A list of corresponding amounts for each ingredient.",
                    "example": [
                      "300g",
                      "2 pcs"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successfully created a recipe",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 201
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully created a recipe!"
                    },
                    "data": {
                      "$ref": "#/components/schemas/recipe"
                    }
                  }
                },
                "examples": {
                  "example": {
                    "summary": "Recipe created",
                    "value": {
                      "status": 201,
                      "message": "Successfully created a recipe!",
                      "data": {
                        "title": "Cake",
                        "category": "Dessert",
                        "instructions": "Follow steps...",
                        "description": "A rich chocolate cake with cream",
                        "thumb": "https://res.cloudinary.com/drxvkqgab/image/upload/v1751815133/azuzlwauesqx75yu7zog.jpg",
                        "time": "60",
                        "cals": 300,
                        "ingredients": [
                          {
                            "id": "640c2dd963a319ea671e367e",
                            "measure": "300g"
                          },
                          {
                            "id": "640c2dd963a319ea671e3687",
                            "measure": "2 pcs"
                          }
                        ],
                        "owner": "685e469e90ec1af79417e1e0",
                        "_id": "686a93dd70a76a8fd13cc748",
                        "createdAt": "2025-07-06T15:18:53.977Z",
                        "updatedAt": "2025-07-06T15:18:53.977Z"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
      "recipe": {
        "type": "object",
        "required": [
          "title",
          "category",
          "instructions",
          "description",
          "ingredients"
        ],
        "properties": {
          "_id": {
            "type": "string",
            "description": "Unique recipe ID (MongoDB ObjectId)",
            "example": "64a1b2c3e25d8c90e88d5678"
          },
          "title": {
            "type": "string",
            "description": "The full name of the recipe",
            "example": "Breakfast Omelette"
          },
          "category": {
            "type": "string",
            "description": "The full name of the category",
            "enum": [
              "Seafood",
              "Lamb",
              "Starter",
              "Chicken",
              "Beef",
              "Dessert",
              "Vegan",
              "Pork",
              "Vegetarian",
              "Miscellaneous",
              "Pasta",
              "Breakfast",
              "Side",
              "Goat",
              "Soup"
            ],
            "example": "Breakfast"
          },
          "area": {
            "type": "string",
            "description": "The cuisine area (e.g. Italian, Mexican)",
            "example": "Italian"
          },
          "instructions": {
            "type": "string",
            "description": "Preparation steps",
            "example": "Crack the eggs and beat with a fork..."
          },
          "description": {
            "type": "string",
            "description": "Short description of the recipe",
            "example": "A light and fluffy omelette perfect for breakfast."
          },
          "thumb": {
            "type": "string",
            "description": "URL to recipe thumbnail image",
            "example": "https://example.com/uploads/image.jpg"
          },
          "time": {
            "type": "string",
            "description": "Time needed to prepare the recipe",
            "example": "30 min"
          },
          "cals": {
            "type": "number",
            "description": "Calories per portion",
            "example": 250
          },
          "ingredients": {
            "type": "array",
            "description": "List of ingredients with their measures",
            "minItems": 1,
            "items": {
              "type": "object",
              "required": [
                "id",
                "measure"
              ],
              "properties": {
                "id": {
                  "type": "string",
                  "description": "Ingredient ID (MongoDB ObjectId, ref to Ingredient)",
                  "example": "64a1b2c3e25d8c90e88d5678"
                },
                "measure": {
                  "type": "string",
                  "description": "Quantity and unit of the ingredient",
                  "example": "2 tbsp"
                }
              }
            }
          },
          "owner": {
            "type": "string",
            "description": "User ID who created the recipe (MongoDB ObjectId)",
            "example": "68419ace9cd45e2f9773af01"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Date of creation",
            "example": "2025-06-28T14:55:00Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Last update date",
            "example": "2025-06-29T10:15:00Z"
          }
        }
      },
...
    },
    "responses": {
...
      "401": {
        "description": "Unauthorized error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message",
                "data"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 401
                },
                "message": {
                  "type": "string",
                  "example": "UnauthorizedError"
                },
                "data": {
                  "type": "object",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Access token expired"
                    }
                  }
                }
              }
            }
          }
        }
      },
...
      "500": {
        "description": "Internal Server error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 500
                },
                "message": {
                  "type": "string",
                  "example": "Internal Server error"
                }
              }
            }
          }
        }
      }
    }
  }
}